refactor: moved user and related child entities to domain group
