+++ src/main/java/se/ltu/eduo/user/repository/UserRepository.java
@@ -1,8 +1,6 @@
-package se.ltu.eduo.repository;
+package se.ltu.eduo.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import se.ltu.eduo.model.User;
-
-import java.util.List;
+import se.ltu.eduo.user.model.User;
 
 public interface UserRepository extends JpaRepository<User, Integer> {}
