+++ src/main/java/se/ltu/eduo/user/request/RegisterRequest.java
@@ -1,3 +1,3 @@
-package se.ltu.eduo.dto.request;
+package se.ltu.eduo.user.request;
 
 public record RegisterRequest(String firstName, String lastName, String username, String password) {}
